+++ packages/sql-openapi/lib/entity-to-routes.js
@@ -3,16 +3,13 @@
 const { mapSQLTypeToOpenAPIType } = require('@platformatic/sql-json-schema-mapper')
 const camelcase = require('camelcase')
 const { singularize } = require('inflected')
-const { generateArgs } = require('./shared')
+const { generateArgs, rootEntityRoutes, capitalize, getFieldsForEntity } = require('./shared')
 
 const getEntityLinksForEntity = (app, entity) => {
   const entityLinks = {}
   for (const relation of entity.relations) {
     const ownField = camelcase(relation.column_name)
     const relatedEntity = app.platformatic.entities[camelcase(singularize(relation.foreign_table_name))]
-    if (relatedEntity.primaryKeys.size !== 1) {
-      continue
-    }
     const relatedEntityPrimaryKeyCamelcase = camelcase(relatedEntity.primaryKeys.values().next().value)
     const relatedEntityPrimaryKeyCamelcaseCapitalized = capitalize(relatedEntityPrimaryKeyCamelcase)
     const getEntityById = `Get${relatedEntity.name}By${relatedEntityPrimaryKeyCamelcaseCapitalized}`
@@ -43,14 +40,6 @@ const getEntityLinksForEntity = (app, entity) => {
   return entityLinks
 }
 
-const getFieldsForEntity = (entity) => ({
-  type: 'array',
-  items: {
-    type: 'string',
-    enum: Object.keys(entity.fields).map((field) => entity.fields[field].camelcase).sort()
-  }
-})
-
 async function entityPlugin (app, opts) {
   const entity = opts.entity
 
@@ -72,90 +61,7 @@ async function entityPlugin (app, opts) {
 
   const fields = getFieldsForEntity(entity)
 
-  app.get('/', {
-    schema: {
-      operationId: 'get' + capitalize(entity.pluralName),
-      querystring: {
-        type: 'object',
-        properties: {
-          limit: { type: 'integer' },
-          offset: { type: 'integer' },
-          totalCount: { type: 'boolean', default: false },
-          fields,
-          ...whereArgs,
-          ...orderByArgs
-        },
-        additionalProperties: false
-      },
-      response: {
-        200: {
-          type: 'array',
-          items: entitySchema
-        }
-      }
-    }
-  }, async function (request, reply) {
-    const query = request.query
-    const { limit, offset, fields } = query
-    const queryKeys = Object.keys(query)
-    const where = {}
-    const orderBy = []
-
-    for (let i = 0; i < queryKeys.length; i++) {
-      const key = queryKeys[i]
-
-      if (key.startsWith('where.')) {
-        const [, field, modifier] = key.split('.')
-        where[field] ||= {}
-        let value = query[key]
-        if (modifier === 'in' || modifier === 'nin') {
-          // TODO handle escaping of ,
-          value = query[key].split(',')
-          if (mapSQLTypeToOpenAPIType(entity.fields[field].sqlType) === 'integer') {
-            value = value.map((v) => parseInt(v))
-          }
-        }
-        where[field][modifier] = value
-      } else if (key.startsWith('orderby.')) {
-        const [, field] = key.split('.')
-        orderBy[field] ||= {}
-        orderBy.push({ field, direction: query[key] })
-      }
-    }
-
-    const ctx = { app: this, reply }
-    const res = await entity.find({ limit, offset, fields, orderBy, where, ctx })
-
-    // X-Total-Count header
-    if (query.totalCount) {
-      let totalCount
-      if ((((offset ?? 0) === 0) || (res.length > 0)) && ((limit === undefined) || (res.length < limit))) {
-        totalCount = (offset ?? 0) + res.length
-      } else {
-        totalCount = await entity.count({ where, ctx })
-      }
-      reply.header('X-Total-Count', totalCount)
-    }
-
-    return res
-  })
-
-  app.post('/', {
-    schema: {
-      body: entitySchema,
-      response: {
-        200: entitySchema
-      }
-    },
-    links: {
-      200: entityLinks
-    }
-  }, async function (request, reply) {
-    const ctx = { app: this, reply }
-    const res = await entity.save({ input: request.body, ctx })
-    reply.header('location', `${app.prefix}/${res.id}`)
-    return res
-  })
+  rootEntityRoutes(app, entity, whereArgs, orderByArgs, entityLinks, entitySchema, fields)
 
   app.get(`/:${primaryKeyCamelcase}`, {
     schema: {
@@ -367,65 +273,6 @@ async function entityPlugin (app, opts) {
     })
   }
 
-  app.put('/', {
-    schema: {
-      body: entitySchema,
-      querystring: {
-        type: 'object',
-        properties: {
-          fields,
-          ...whereArgs
-        },
-        additionalProperties: false
-      },
-      response: {
-        200: {
-          type: 'array',
-          items: entitySchema
-        }
-      }
-    },
-    links: {
-      200: entityLinks
-    },
-    async handler (request, reply) {
-      const ctx = { app: this, reply }
-      const query = request.query
-      const queryKeys = Object.keys(query)
-      const where = {}
-
-      for (let i = 0; i < queryKeys.length; i++) {
-        const key = queryKeys[i]
-        if (key.startsWith('where.')) {
-          const [, field, modifier] = key.split('.')
-          where[field] ||= {}
-          let value = query[key]
-          if (modifier === 'in' || modifier === 'nin') {
-            // TODO handle escaping of ,
-            value = query[key].split(',')
-            if (mapSQLTypeToOpenAPIType(entity.fields[field].sqlType) === 'integer') {
-              value = value.map((v) => parseInt(v))
-            }
-          }
-          where[field][modifier] = value
-        }
-      }
-
-      const res = await entity.updateMany({
-        input: {
-          ...request.body
-        },
-        where,
-        fields: request.query.fields,
-        ctx
-      })
-      // TODO: Should find a way to test this line
-      // if (!res) return reply.callNotFound()
-      reply.header('location', `${app.prefix}`)
-      return res
-    }
-  })
-
   app.delete(`/:${primaryKeyCamelcase}`, {
     schema: {
       params: primaryKeyParams,
@@ -471,10 +318,6 @@ function getPrimaryKeyParams (entity) {
     properties,
     required
   }
-}
-
-function capitalize (str) {
-  return str.charAt(0).toUpperCase() + str.slice(1)
 }
 
 module.exports = entityPlugin
